requirements.txt
streamlit==1.32.0
transformers==4.38.1
torch==2.2.1
accelerate==0.25.0
bitsandbytes==0.41.3
llama-index==0.9.11
llama-index-embeddings-huggingface==0.1.4
chromadb==0.4.22
python-telegram-bot==20.7
python-dotenv==1.0.0
Pillow==10.1.0
huggingface-hub==0.20.3
sentencepiece==0.1.99
safetensors==0.4.2
einops==0.7.0
langchain==0.0.352
langchain-community==0.0.16
zipfile36==0.1.3
nest_asyncio==1.6.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.26.0